Rename `src/operations` to `src/operation` and move `AbstractOperation` in this folder
